Major Updates. All features are working.

+++ src/main/java/com/task/e_commerce/dtos/ProductDto.java
@@ -13,8 +13,8 @@ public class ProductDto {
     private Long productCode;
     private String name;
     private String composition;
-    private Long mrp;
-    private Long salesRate;
+    private Double mrp;
+    private Double salesRate;
     private Long totalStrip;
     private Long medicinePerStrip;
     private String imageUrl;
